backend/requirements.txt
# AdaptPath Backend Dependencies - VERSION COMPATIBLE
# Install with: pip install -r requirements.txt

# FastAPI Framework with compatible Pydantic version
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Pydantic - Compatible version
pydantic==2.9.2
pydantic-settings==2.6.1

# Google Generative AI (NEW package)
google-genai==0.3.0

# Environment variables
python-dotenv==1.0.1

# Date and time handling
python-dateutil==2.9.0

# Async support
httpx==0.27.2
aiofiles==24.1.0

# Type hints
typing-extensions==4.12.2